requirements.txt
google-generativeai
pydantic-ai
fastapi